avoid repeating for child module and refactor code
