+++ main/plugins/org.talend.repository/src/main/java/org/talend/repository/ui/utils/UpdateLog4jJarUtils.java
@@ -2,6 +2,8 @@ package org.talend.repository.ui.utils;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
+import java.util.Comparator;
 import java.util.Iterator;
 import java.util.LinkedHashSet;
 import java.util.List;
@@ -308,5 +310,71 @@ public class UpdateLog4jJarUtils {
             }
         }
         return false;
+    }
+
+    public static void sortClassPath4Log4j(Set<ModuleNeeded> highPriorityModuleNeeded, List<ModuleNeeded> neededModules) {
+        Collections.sort(neededModules, new Comparator<ModuleNeeded>() {
+
+            @Override
+            public int compare(ModuleNeeded o1, ModuleNeeded o2) {
+                if (highPriorityModuleNeeded == null) {
+
+                    return 0;
+                }
+                for (String moduleName : MODULES_NEED_ADDED_BACK) {
+                    if (StringUtils.equals(moduleName, o1.getModuleName())
+                            && !StringUtils.equals(moduleName, o2.getModuleName())) {
+                        return -1;
+                    }
+                    if (!StringUtils.equals(moduleName, o1.getModuleName())
+                            && StringUtils.equals(moduleName, o2.getModuleName())) {
+                        return 1;
+                    }
+                }
+                return 0;
+
+            }
+        });
+
+        Collections.sort(neededModules, new Comparator<ModuleNeeded>() {
+
+            @Override
+            public int compare(ModuleNeeded o1, ModuleNeeded o2) {
+                for (String moduleName : MODULES_NEED_UPDATE_ORDER) {
+                    if (StringUtils.equals(moduleName, o1.getModuleName())
+                            && !StringUtils.equals(moduleName, o2.getModuleName())) {
+                        return -1;
+                    }
+                    if (!StringUtils.equals(moduleName, o1.getModuleName())
+                            && StringUtils.equals(moduleName, o2.getModuleName())) {
+                        return 1;
+                    }
+                }
+                return 0;
+
+            }
+        });
+
+        Collections.sort(neededModules, new Comparator<ModuleNeeded>() {
+
+            @Override
+            public int compare(ModuleNeeded o1, ModuleNeeded o2) {
+                if (highPriorityModuleNeeded == null) {
+                    return 0;
+                }
+                for (ModuleNeeded module : highPriorityModuleNeeded) {
+                    if (StringUtils.equals(module.getModuleName(), o1.getModuleName())
+                            && !StringUtils.equals(module.getModuleName(), o2.getModuleName())) {
+                        return -1;
+                    }
+                    if (!StringUtils.equals(module.getModuleName(), o1.getModuleName())
+                            && StringUtils.equals(module.getModuleName(), o2.getModuleName())) {
+                        return 1;
+                    }
+                }
+                return 0;
+
+            }
+        });
     }
 }
